% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{object of class formula to fit a linear model for.}

\item{data}{input data frame.}
}
\value{
A list with elements \code{Coefficients} which is a matrix containing
coefficients, standard error, t value, and p value, \code{ybar} the predictions
}
\description{
This function fits a linear model.
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)
my_lm(lifeExp ~ gdpPerCap + continent, data = my_gapminder)

}
