% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Linear Model Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{integer representing number of folds.}
}
\value{
A numeric with the cross validation error.
}
\description{
This function fits a linear model.
}
