% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test Function}
\usage{
my_t.test(x, alternative = "two.sided", mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{character string specifying the alternative hypothesis
and this should be \code{"two.sided"}, \code{"less"}, or \code{"greater"}
otherwise an error will be thrown. Default is \code{"two.sided"}.}

\item{mu}{a numeric indicating the null hypothesis value of the mean.}
}
\value{
A list with elements: \code{test_stat} the numeric test
statistic, \code{df} the degrees of freedom, \code{alternative} the value of
the parameter \code{alternative}, and \code{p_val} the numeric p value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(my_gapminder$lifeExp, alternative = "less", 60)
my_t.test(my_gapminder$lifeExp, alternative = "greater", 60)
my_t.test(my_gapminder$lifeExp, 60)

}
