% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{object of class formula to fit a linear model for.}

\item{data}{input data frame.}
}
\value{
A matrix summarizing the coefficients of the linear fit.
}
\description{
This function fits a linear model.
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars)

}
